# docs/swagger.yaml
basePath: /v1
definitions:
...
  internal_api.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  internal_api.ScriptRevisionResponse:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      description:
        type: string
      id:
        type: string
      loadTestId:
        type: string
      revisionNumber:
        type: integer
      scriptContent:
        description: Base64 encoded Python script
        type: string
    type: object
...
  internal_api.UpdateScriptRequest:
    properties:
      description:
        description: Optional change description
        type: string
      scriptContent:
        description: Base64 encoded Python script
        type: string
      updatedBy:
        type: string
    required:
    - scriptContent
    - updatedBy
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@loadmanager.io
    name: API Support
  description: |-
    Load testing management platform for creating, executing, and monitoring distributed load tests with Locust.

    ## Features
    - **Load Test Management**: Create and manage load tests with script versioning
    - **Script Revisions**: Full version control for test scripts with audit trail
    - **Test Execution**: Start and monitor load test runs with real-time metrics
    - **Visualization**: Real-time dashboards and historical metrics

    ## Authentication
    Currently, this API does not require authentication. Future versions will include API key or OAuth2 support.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Load Manager API
  version: "1.0"
paths:
...
  /load-tests/{id}/script:
    get:
      description: Returns the latest script revision for the load test
      parameters:
      - description: Load Test ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Latest script revision
          schema:
            $ref: '#/definitions/internal_api.ScriptRevisionResponse'
        "404":
          description: Script not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Get latest script
      tags:
      - Scripts
    put:
      consumes:
      - application/json
      description: Creates a new script revision for the load test with base64 encoded
        content
      parameters:
      - description: Load Test ID
        in: path
        name: id
        required: true
        type: string
      - description: Script content (base64) and description
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.UpdateScriptRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Script revision created successfully
          schema:
            $ref: '#/definitions/internal_api.ScriptRevisionResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: Load test not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Failed to create script revision
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Update load test script
      tags:
      - Scripts
  /load-tests/{id}/script/revisions:
    get:
      description: Returns all script revisions for a load test (most recent first)
      parameters:
      - description: Load Test ID
        in: path
        name: id
        required: true
        type: string
      - default: 10
        description: Maximum number of revisions to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of script revisions
          schema:
            items:
              $ref: '#/definitions/internal_api.ScriptRevisionResponse'
            type: array
        "500":
          description: Failed to list script revisions
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: List script revision history
      tags:
      - Scripts
  /load-tests/{id}/script/revisions/{revisionId}:
    get:
      description: Returns a specific script revision by its ID
      parameters:
      - description: Load Test ID
        in: path
        name: id
        required: true
        type: string
      - description: Revision ID
        in: path
        name: revisionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Script revision details
          schema:
            $ref: '#/definitions/internal_api.ScriptRevisionResponse'
        "404":
          description: Script revision not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Get specific script revision
      tags:
      - Scripts
...
schemes:
- http
- https
swagger: "2.0"
...